requirements.txt
django==3.2.12
djangorestframework==3.11.0
markdown==3.2.2
django-filter==2.2.0
channels==2.4.0
channels_redis==3.2.0
gunicorn==20.0.4
django-cors-headers==3.7.0
python-dateutil==2.8.1
pandas==1.1.5
django-cookies-samesite==0.8.0
mysqlclient==2.0.3
ip2geotools==0.1.5
pymysql==1.0.2
sqlalchemy==1.3.19
faker==8.4.0
redis==3.5.3
pytest-django==4.5.2
celery==5.1.2
pytest-remotedata==0.3.3
